requirements.txt
numpy==2.2.6
pillow==11.3.0
tensorflow==2.20.0
tensorflow-datasets
